refactor: support 64-bit immediates and variable-length operands

+++ compiler.h
@@ -21,7 +21,7 @@ typedef struct {
 	TokenType type;
 	const char *start;
 	unsigned int length;
-	int32_t value; /* if tok == TOK_INT */
+	int64_t value; /* if tok == TOK_INT */
 } Token;
 
 typedef struct {
